+++ src/__tests__/vizzu-player._setSlides.test.js
@@ -2,7 +2,7 @@ import VizzuPlayer from "../vizzu-player.js";
 
 import { slides } from "./assets/slides.js";
 
-import VizzuMock from "../__mocks__/vizzu.js";
+import VizzuMock from "./assets/mocks/vizzu.js";
 
 global.window.Vizzu = VizzuMock;
